refactor: move manageableNetworkTokens to network-tokens.derived.ts
