+++ src/frontend/src/lib/derived/network.derived.ts
@@ -1,6 +1,4 @@
 import { icrcAccountIdentifierText } from '$icp/derived/ic.derived';
-import { buildIcrcCustomTokens } from '$icp/services/icrc-custom-tokens.services';
-import type { IcrcCustomToken } from '$icp/types/icrc-custom-token';
 import { DEFAULT_NETWORK, DEFAULT_NETWORK_ID } from '$lib/constants/networks.constants';
 import { address } from '$lib/derived/address.derived';
 import { routeNetwork } from '$lib/derived/nav.derived';
@@ -36,21 +34,4 @@ export const networkAddress: Readable<OptionAddress | string> = derived(
 	[address, icrcAccountIdentifierText, networkICP],
 	([$address, $icrcAccountIdentifierStore, $networkICP]) =>
 		$networkICP ? $icrcAccountIdentifierStore : $address
-);
-
-export const manageableNetworkTokens: Readable<ManageableToken[]> = derived(
-	[networkTokens],
-	([$networkTokens]) => {
-		const allIcrcCustomTokens: IcrcCustomToken[] = buildIcrcCustomTokens()
-			.map((token) => ({
-				...token,
-				id: Symbol(token.symbol),
-				enabled: false
-			}))
-			.filter((token) => token.indexCanisterVersion !== 'outdated');
-		return mergeTokenLists<ManageableToken>(
-			$networkTokens.map((token) => ({ ...token, enabled: true })),
-			allIcrcCustomTokens
-		);
-	}
 );
